screenshot/requirements.txt
python-dotenv
selenium
Pillow
webdriver-manager
requests
beautifulsoup4
